chore: remove PR templates folder
Using a folder is good for multiple templates but would demand
additional configuration. This is the reason it is not being selected
for new PRs. Since the most common case is to have one single template
for PRs, it was moved to .github to make it work in a simpler way.

Signed-off-by: Marcus Burghardt <maburgha@redhat.com>