feat: Add plugin configuration support
Common logic for plugin launch and configuration are centralized
in the complytime package. A PluginOptions struct with custom struct
tags was chosen to represent complytime fixed, known global options to support
validation.

The C2P plugin manager accepts a dynamic map of option selections for
each plugin so the custom struct tags are used to convert the struct to the
map form. Another could be to use json tags and json encode/decodings, but
this requires some intermediate processing and using reflection in this case
was straightforward because the struct is simple.

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>