chore: tailoring ignore already unselect rules
During tests it was noticed a particular case with ANSSI profile in a
Datastream where some groups are explicitily unselected. Since these
groups are also not part of the OSCAL profile they were included in the
tailoring file even if already unselected in Datastream. This does not
cause issues but creates unnecessary content in tailoring file.

Signed-off-by: Marcus Burghardt <maburgha@redhat.com>