feat: discover datastream based on system information
Instead of requiring the Datastream to be manually informed via
configuration file, the option was set to an empty value as default and
the plugin will try to dynamically determine the correct Datastream
based on system information. In exceptional cases, the Datastream can
still be informed via configuration file. SanitizeAndValidatePath
function was also removed since it is no longer necessary, so better to
remove it sooner than later.

Signed-off-by: Marcus Burghardt <maburgha@redhat.com>